python-backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
tensorflow==2.13.0
pillow==10.0.1
gtts==2.3.2
pyttsx3==2.90